+++ bin/replicante/src/components/webui/constants.rs
@@ -1,0 +1,2 @@
+pub const FIND_CLUSTERS_LIMIT: u8 = 25;
+pub const RECENT_EVENTS_LIMIT: i64 = 100;