+++ CHANGELOG.md
@@ -6,14 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- HTTP Discovery
+- Cluster specific events API (`/cluster/:id/events`).
+- HTTP Discovery.
 
 ### Changed
 - **BREAKING**: Change `storage` configuration to `storage.primary`.
 - **BREAKING**: Store every event in the "view" database.
 
 ### Removed
-- **BREAKING**: File Discovery
+- **BREAKING**: File Discovery.
 
 ## [0.3.1] - 2019-07-04
 ### Added